% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10.DPIROC.R
\name{dpi_ps_roc}
\alias{dpi_ps_roc}
\title{DPI PS Score ROC Analysis}
\usage{
dpi_ps_roc(dpi_base, ttd_path = NULL, drugbank_path = NULL)
}
\arguments{
\item{dpi_base}{Character. Base directory in which to search for dpi_ps_score.txt files (recursively).}

\item{ttd_path}{Character. File path to the DTI_TTD.RData file (loads object TTD). If NULL, defaults to file.path(dpi_base, "DTI_TTD.RData").}

\item{drugbank_path}{Character. File path to the DTI_DrugBank.RData file
(loads object \code{DrugBank}). If NULL, defaults to file.path(dpi_base, "DTI_DrugBank.RData").}
}
\value{
ROC_<dpi_ps_score.txt> file in the same directory containing cutoff, AUC, drug count, and DPI count.
}
\description{
Perform ROC analysis on all dpi_ps_score.txt files under a specified directory, using known drug–target data to evaluate prediction performance across score cutoffs.
}
\details{
DETAILS
}
\examples{
\dontrun{
dpi_ps_roc(
    dpi_base      = "./PRS/PS",
    ttd_path      = "./PRS/PS/DTI_TTD.RData",
    drugbank_path = "./PRS/PS/DTI_DrugBank.RData"
}
}
