% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.DrugResponse.R
\name{drug_response_analysis}
\alias{drug_response_analysis}
\title{Drug Response Analysis for Subtype-Specific Gene Modules}
\usage{
drug_response_analysis(
  base_path,
  subtype_file,
  expr_base_path,
  expr_data_file,
  diff_expr_file,
  module_base_path,
  drug_response_path,
  gdsc_expr_file,
  gdsc_res_file,
  network_methods,
  module_methods,
  gene_col_name
)
}
\arguments{
\item{base_path}{Character string specifying the base directory for all input/output paths.}

\item{subtype_file}{Character string specifying the path to the subtype phenotype data file.}

\item{expr_base_path}{Character string specifying the directory for expression data output.}

\item{expr_data_file}{Character string specifying the path to the expression data file.}

\item{diff_expr_file}{Character string specifying the path to the differential expression results file.}

\item{module_base_path}{Character string specifying the directory containing module selection files.}

\item{drug_response_path}{Character string specifying the directory for drug response output.}

\item{gdsc_expr_file}{Character string specifying the path to the GDSC expression data file.}

\item{gdsc_res_file}{Character string specifying the path to the GDSC response data file.}

\item{network_methods}{Character vector specifying the PPI network databases (e.g., "String", "physicalPPIN", "chengF").}

\item{module_methods}{Character vector specifying the module detection methods (e.g., "Louvain", "WF").}

\item{gene_col_name}{Character string specifying the column name for genes in expression data.}
}
\value{
Saves drug response predictions, clustering results, and drug response networks (DRNs) to the specified output directory.
}
\description{
This function performs drug response analysis using oncoPredict and ConsensusClusterPlus  to predict drug responses and cluster samples based on module-specific gene expression.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 drug_response_analysis(
 base_path = "your_workspace",
 subtype_file = "subtype.xlsx",
 expr_base_path = "DrugResponse/subtype_expression",
 expr_data_file = "expression.xlsx",
 diff_expr_file = "DEPS_result/diff_expression_results_all.xlsx",
 module_base_path = "ModuleSelection",
 drug_response_path = "DrugResponse",
 gdsc_expr_file = "GDSC2_Expr.RData",
 gdsc_res_file = "GDSC2_Res.RData",
 network_methods = c("String", "physicalPPIN", "chengF"),
 module_methods = c("Louvain", "WF"),
 gene_col_name = "Gene")
 }
}
}
