% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.ModuleAnnotation.R
\name{enrich.GO}
\alias{enrich.GO}
\title{Perform GO Enrichment Analysis}
\usage{
enrich.GO(
  gene,
  ont = "all",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.2,
  universe = NULL,
  minGSSize = 10,
  maxGSSize = 500
)
}
\arguments{
\item{gene}{Character vector of gene symbols for analysis.}

\item{universe}{Character vector of background genes (default: NULL).}

\item{minGSSize}{Minimum gene set size (default: 10).}

\item{maxGSSize}{Maximum gene set size (default: 500).}

\item{Character}{string specifying p-value adjustment method (default: "BH").}

\item{Numeric}{threshold for q-value (default: 0.2).}
}
\value{
Data frame of enrichment results or NULL if no significant results.
}
\description{
Conducts GO enrichment analysis using MSigDB gene sets for specified ontology.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
