% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.ModuleAnnotation.R
\name{enrich.Plot}
\alias{enrich.Plot}
\title{Plot Enrichment Results}
\usage{
enrich.Plot(plotData, saveName)
}
\arguments{
\item{plotData}{Data frame containing enrichment results with required columns (Category, Description, p.adjust, Count, geneRatio).}

\item{saveName}{Character string specifying the output file path for the plot.}
}
\value{
Saves a dot plot to the specified file path.
}
\description{
Generates a dot plot for enrichment results and saves it as a PDF.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[egg]{set_panel_size}}
}
