% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.ModuleAnnotation.R
\name{functional_annotation}
\alias{functional_annotation}
\title{Perform Functional Annotation}
\usage{
functional_annotation(
  subtype_file = "./subtype.xlsx",
  base_input_path = "./ModuleSelection/",
  base_output_path = "./FunctionalAnnotation/",
  network_method = c("String", "physicalPPIN", "chengF"),
  module_method = c("Louvain", "WF")
)
}
\arguments{
\item{subtype_file}{Character string specifying the path to the subtype phenotype data file (default: "./subtype.xlsx").}

\item{base_input_path}{Character string specifying the base directory containing module selection files (default: "./ModuleSelection/").}

\item{base_output_path}{Character string specifying the base directory for output files (default: "./FunctionalAnnotation/").}

\item{module_method}{Character vector specifying the module detection methods (options: "Louvain", "WF").}

\item{network_methodCharacter}{vector specifying the PPI network databases (options: "String", "physicalPPIN", "chengF").}
}
\value{
Saves enrichment results and plots to the specified output directory.
}
\description{
Performs GO, KEGG, and Hallmark enrichment analyses for gene modules and saves results.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
