% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.SEQCre.R
\name{run_SEQCre}
\alias{run_SEQCre}
\title{Generate Protein and Drug Sequence Files for Drug Response Prediction}
\usage{
run_SEQCre(input_base, output_base, subtype_file, py_env = "py3.9")
}
\arguments{
\item{input_base}{Character. Path to directory containing Drug Response Network (DRN) input files. Each file should be named in the format "*_DRN_info.txt".}

\item{output_base}{Character. Path to directory where output sequence files will be saved.}

\item{subtype_file}{Character. Path to Excel file containing subtype phenotype data.}

\item{py_env}{Character. Name or path of Conda environment containing required Python  dependencies (default: "py3.9").}
}
\value{
Saves sequence output directory.
}
\description{
Processes drug response network data to generate protein sequence (FASTA) and
drug SMILES files required for DeepPurpose-based drug-target interaction prediction.
This function serves as an R wrapper for Python sequence processing functionality.
}
\details{
DETAILS
}
\examples{
\dontrun{
run_SEQCre(
  input_base = "data/DRN_files",
  output_base = "output/sequence_files",
  subtype_file = "metadata/subtype_info.xlsx",
  py_env = "drug_discovery"
)
}
}
