% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.BindingScore.R
\name{predict_drug_response}
\alias{predict_drug_response}
\title{Drug Response Prediction Using DeepPurpose}
\usage{
predict_drug_response(
  subtype_file,
  drn_base,
  prs_base,
  output_base,
  py_env,
  network_methods,
  module_methods
)
}
\arguments{
\item{subtype_file}{Character string specifying the path to the subtype phenotype data file.}

\item{drn_base}{Character string specifying the base directory containing Drug Response Network (DRN) files.}

\item{prs_base}{Character string specifying the base directory containing Protein and Drug sequence files.}

\item{output_base}{Character string specifying the base directory for output files.}

\item{py_env}{Character string specifying the path to the Python Conda environment with DeepPurpose installed.}

\item{network_methods}{Character vector specifying the PPI network databases (e.g., "string", "physicalppin", "chengf").}

\item{module_methods}{Character vector specifying the module detection methods (e.g., "Louvain", "WF").}
}
\value{
Saves DPI information and drug-target interaction predictions to the specified output directory.
}
\description{
This function predicts drug-target interactions using the DeepPurpose Python library for specified subtypes, network methods, and module methods, integrating DRN and SEQ data.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 predict_drug_response(
 subtype_file = "./subtype.xlsx",
 drn_base = "./DrugResponse",
 prs_base = "./PRS/SMILES_SEQ",
 output_base = "./PRS/PS",
 py_env = "./envs/py3.9",
 network_methods = c("string", "physicalppin", "chengf"),
 module_methods = c("Louvain", "WF")
 )
 }
}
}
