% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.PScore.R
\name{process_prs_dti}
\alias{process_prs_dti}
\title{Process PRS-DTI Data}
\usage{
process_prs_dti(
  subtype_file,
  prs_ps_base,
  network_methods = c("string", "physicalppin", "chengf"),
  module_methods = c("Louvain", "WF"),
  py_env
)
}
\arguments{
\item{subtype_file}{Character string specifying the path to the subtype phenotype data file.}

\item{prs_ps_base}{Character string specifying the base directory containing PRS and DTI files.}

\item{network_methods}{Character vector specifying the PPI network databases (default: c("string", "physicalppin", "chengf")).}

\item{module_methods}{Character vector specifying the module detection methods (default: c("Louvain", "WF")).}

\item{py_env}{Character string specifying the path to the Python Conda environment with DeepPurpose installed.}
}
\value{
Saves PRS-DTI scores and related files to the specified output directories.
}
\description{
Perturbation Response Scanning (PRS) and Drug-Target Interaction (DTI) data using DeepPurpose and compute protein-drug interaction scores for each subtype, network, and module.
}
\details{
DETAILS
}
\examples{
\dontrun{
 process_prs_dti(
    subtype_file    = "./subtype.xlsx",
    prs_ps_base     = "./PRS/PS",
    network_methods = c("string","physicalppin","chengf"),
    module_methods  = c("Louvain","WF"),
    py_env          = "./envs/py3.9"
 )
 }
}
