% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_subnetDR.R
\name{run_subnetDR_pipeline}
\alias{run_subnetDR_pipeline}
\title{Run SubnetDR Pipeline}
\usage{
run_subnetDR_pipeline(base_path, py_env)
}
\arguments{
\item{base_path}{Character string specifying the base directory path containing input files and where output files will be saved.}

\item{py_env}{Character string specifying the path to the Python environment used for certain pipeline steps (e.g., drug response prediction).}
}
\value{
The function generates various output files in the specified directories and prints a completion message.
}
\description{
Executes the complete SubnetDR pipeline for differential expression analysis, network construction, module division, selection, functional annotation, enrichment plotting, drug response analysis, and drug-target interaction processing.
}
\details{
details
}
\examples{
\dontrun{
run_subnetDR_pipeline(
  base_path = "F:/subnetDR/sample_test",
  py_env = "C:/Users/YangMiao/.conda/envs/py3.9")
}
}
