% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.PScore.R
\name{split_edge_module}
\alias{split_edge_module}
\title{Split Edge Module Data}
\usage{
split_edge_module(input_file, out_base)
}
\arguments{
\item{input_file}{Character string specifying the path to the input edge module file.}

\item{out_base}{Character string specifying the base directory for output files.}
}
\value{
Saves edge module files for each valid module to the specified output directory; returns NULL if no valid modules.
}
\description{
Splits edge module data into separate files based on module numbers and saves them to specified directories.
}
\details{
DETAILS
}
\examples{
\dontrun{
files <- list.files(
    path       = "./ModuleDivision",
    pattern    = "^edge_Module_.*\\\\.txt$",
    recursive  = TRUE,
    full.names = TRUE
)
purrr::walk(files, split_edge_module,out_base)
}
}
\seealso{
\code{\link[purrr]{map}}
}
